修复配方异常，修复觉醒双生数据异常，调整战利品表，减少技能爆炸范围为

+++ src/main/java/hua/huase/shanhaicontinent/capability/playerattribute/PlayerAttributeCapability.java
@@ -140,7 +140,7 @@ public class PlayerAttributeCapability extends CapabilityAttributeBase implement
         return maxjingyan;
     }
 
-    public void setMaxjingyan(int maxjingyan) {
+    public void setMaxjingyan(float maxjingyan) {
         this.maxjingyan = maxjingyan;
     }
